Add student form to add students to database

+++ app/Http/Controllers/StudentController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Student;
+use App\Models\Classroom;
+use App\Models\User;
+use App\Models\Gender;
+use App\Models\Nationality;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rules\Password;
 
@@ -13,12 +16,22 @@ class StudentController extends Controller
      */
     public function index()
     {
-        $students = Student::paginate(15);
+        $students = User::role('student')->paginate(15);
         return view('admin.students', ['students' => $students]);
         // return response()->json([
         //     'success' => true,
         //     'students' => Student::all(),
         // ]);
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     */
+    public function create()
+    {
+        $genders = Gender::all();
+        $nationalities = Nationality::all();
+        return view('partials.student-form', ['genders' => $genders, 'nationalities' => $nationalities]);
     }
 
     /**
@@ -29,23 +42,30 @@ class StudentController extends Controller
         $validated = $request->validate([
             'first_name' => 'required|string|max:20',
             'last_name' => 'required|string|max:20',
-            "father's_name" => 'required|string|max:30',
-            "mother's_name" => 'required|string|max:30',
-            'phone_number' => 'required|string|max:10',
+            'father_name' => 'required|string|max:30',
+            'mother_name' => 'required|string|max:30',
+            'phone_number' => 'required|string|unique:users,phone_number|max:10',
             'address' => 'required|string|max:20',
             'date_of_birth' => 'required|date',
-            'email' => 'required|email',
+            'email' => 'required|email|unique:users,email',
             'password' => Password::min(8)
                             ->letters()
                             ->mixedCase()
                             ->numbers()
-                            ->symbols()
-                            ->uncompromised(3),
-            'gender_id' => 'required|integer|max_digits:1',
-            'nationality_id' => 'required|integer|max_digits:2',
+                            ->symbols(),
+            'gender_id' => 'required|integer|exists:genders,id',
+            'nationality_id' => 'required|integer|exists:nationalities,id',
+            // 'classroom_id' => 'required|integer|exists:classroom,id',
         ]);
         $validated['password'] = bcrypt($validated['password']);
-        $student = Student::create($validated);
+        $student = User::create($validated);
+        $student->assignRole('student');
+        // $class = Classroom::find($validated['classroom_id']);
+        // if($class->number_of_students < $class->max_number_of_students) {
+        //     $class->update([
+        //         'number_of_students' => $class->number_of_students + 1,
+        //     ]);
+        // }
         return response()->json([
             'success' => true,
             'message' => 'Student created successfully',
@@ -56,7 +76,7 @@ class StudentController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Student $student)
+    public function show(User $student)
     {
         return response()->json([
             'success' => true,
@@ -67,7 +87,7 @@ class StudentController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Student $student)
+    public function update(Request $request, User $student)
     {
         $validated = $request->validate([
             'first_name' => 'required|string|max:20',
@@ -89,6 +109,7 @@ class StudentController extends Controller
         ]);
         $validated['password'] = bcrypt($validated['password']);
         $student->update($validated);
+        $student->assignRole('student');
         return response()->json([
             'success' => true,
             'message' => 'Student updated successfully',
@@ -99,12 +120,19 @@ class StudentController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Student $student)
+    public function destroy(User $student)
     {
-        $student->delete();
-        return response()->json([
-            'success' => true,
-            'message' => 'Student deleted successfully',
-        ]);
+        if(User::role('student')) {
+            $student->delete();
+            return response()->json([
+                'success' => true,
+                'message' => 'Student deleted successfully',
+            ]);
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'The deleted user is not a student',
+            ]);
+        }
     }
 }
